Add a /compare page for testing / debugging model data.
This is an experimental page for internal usage but needs to live in the
website so we can share all of the chart loading / rendering code with the rest
of the website.

The idea is to let us easily compare all of the (state) graphs for e.g. staging
vs. prod or a modeler can point it at their local model files and compare them
as well.

I think this will be useful, but for now it's experimental.  We can iterate and
or abandon it depending on future needs.

One specific improvement I'd like to make is adding logic to do a simple
algorithmic comparison of each model and then sort them based on how different
they seem to be.