fix: interpolation of overwhelmed date
#86 this doesn't apply to changing the flag based on the selected
series, but does fix the positioning of the flag.

the original interpolation wasn't placing the line close to the
intersection. It seems like it was using a portion of the bed shortfall
as a stand in for a duration, which doesn't make much sense to me.

the new appoaches finds the beginning and end of the window in
which the hospitalization and beds lines first intersect. finds the
intersection point, then uses that to interpolate the day range.