Remove use of county_summaries/XX.summary.json files in the frontend.
We used to need these files to render the county map without having to read all
the projection files. But the calculated_county_interventions.json file took
the place of that a while ago so it's been redundant for a while.

Additionally, the summary files are slightly wrong. E.g. they indicate Weber
County, UT (fips 49057) should have data, but it does not. I'm not sure why
the summary file is wrong, but we don't need it anymore, so who cares.