Use actuals where available in ICU Headroom metric; update AZ and DC decomp. (#1011)
This uses ICUBeds.currentUsageCovid and/or ICUBeds.currentUsageTotal when available in the API to calculate the ICU Headroom metric. Since either one or both may be available, there are 4 possible combinations, and this code handles all of them and subtly includes the result in the copy to make it possible to decode the calculation.

1. If neither are available, you get the existing site behavior where both covid and non-covid patients are estimates:
2. If both are available, then we omit "we estimate" in both places:
3. If only ICUBeds.currentUsageCovid is available, then non-covid patients is an estimate and we include the typical utilization + decomp percentage that we're using to estimate it.
4. If only ICUBeds.currentUsageTotal is available, then both covid and non-covid patients are estimates but we omit the percentage for non-covid patients, since we can avoid using the typical utilization+decomp percentage.

I also updated AZ decomp to 40% to better match up with actuals from their dashboard, as well as DC decomp to 15% to match up with their published numbers.