Fix inconsistent tracking of switch component clicks due to extra event callback. (#2002)
SwitchComponent had two separate callbacks (onChange and onGridClick) and we were only tracking analytics events when you hit the onChange handler (concretely, if you clicked the actual switch UI then onChange would fire and we'd track the analytics for it; but if you clicked on the "States" or "Counties" labels, then onGridClick would fire instead and we wouldn't track it). I reworked it to only have a single event callback and also added it to storybook and did some minor cleanup.