Update alert/newsleter signup to not store alert subscribers in Campaign Monitor.
For historical reasons, when somebody signed up for alerts or the newsletter, we were using a
Campaign Monitor signup form that would:
1. Always add the user to the "Alert Signups" list in Campaign Monitor.
2. Optionally (depending on whether they checked the newsletter checkbox) add them to the "Full Sign
Up List_Scrubbed" list in Campaign Monitor.

But we don't actually use Campaign Monitor at all for alert emails anymore, so we don't need to sign
them up to the first list, and it was costing us extra money.

So I've regenerated the form so that it only signs them up for "Full Sign Up List_Scrubbed" and we
only actually submit the form if they check the newsletter checkbox.

Since we only *sometimes* submit the Campaign Monitor signup form now, I had to invent a new way to
confirm the subscription if they sign up for alerts but not the newsletter, so I added a little UI
to change the "Sign up" button to say "Subscribed!" for 3 seconds after