Fix email alert signups report. (#2708)
There was an off-by-1 error with the "subscriptions by date" tracking that I've fixed.

Also, the script was silently failing every day due to too many cells in the spreadsheet. I deleted extra rows/columns, which seems to have fixed it, but I also changed the script so that it will fail more loudly in the future.