Homepage map code cleanup (#3825)
This is mostly refactoring with the goal of deleting dead code, cleaning up some naming, and making the USMap component accept getFillColor and renderTooltip properties so that we can have USRiskMap and (in #3826) USVaccineMap components that customize the coloring / tooltips. But there is one minor functionality change: I added an onClick handler to our entire Embed component so that clicking anywhere on the embed will take you to our homepage. This let me remove the onClick handler on the map that wasn't used anywhere else.

* Remove unused MiniMap.
* Remove unused MapInstructions.
* Remove unused LegendTitle.
* Move Legend to SocialLocationPreview component since it's the only place that needs it.
* Remove onClick handler from Map and move it into Embed, the only place it's used.
* Bunch of minor renaming.
  * src/components/Map => src/components/USMap
  * USACountyMap => USMap
  * Map => USRiskMap
* Add getFillColor to USMap - Move fill color calculation out of USMap and into USRiskMap.
* Add renderTooltip property to USMap - Move tooltip rendering out of USMap and into USRiskMap.