Simplify / fix VaccineProgressBar sizing.
I realized that by adding `width: 100%` to the svg CSS styles, it actually negated the `width` property on the svg element itself, which was being set by the responsive ParentSize wrapper, and so in fact we didn't need that wrapper at all, and removing it makes the progress bar render immediately (the flicker problem that required me to add `width: 100%` in the first place...

I ended up just removing the `width` property entirely, which meant I had to specify it on the container in VaccinationsTable, but other than that everything still works and with less code. So I think this is a win?