Fix NavBar grid so that back arrow, etc. work on Chrome 93.
We accidentally set up our grid columns so that the search bar grid cell
overlapped the back arrow and logo (and a little bit of the donate button),
which apparently worked fine until Chrome 93. But with Chrome 93, the search
bar div ended up on top of the other elements and e.g. you couldn't click the back
arrow / logo anymore.