Remove legacy dead code from SummaryStats component.
SummaryStats used to be used for the LocationPage above-the-fold, but we rewrote it a long time ago so now it's only used by Embed and SocialLocationPreview. So it contained a lot of dead code.

This also lets us get rid of the detail() method from LevelInfo, which cleans up some legacy metric code.

I also deleted the unused SignalStatus component which used to be used by SummaryStats but its usage was removed a while ago.

Remove dead SignalStatus component.