Update region-overrides.json
There is a cross repo dependency for this file. The latest version (this applied patch) was shipped to the downstream repo yesterday, but was accidentally clobbered this morning by a GH action that resolves these dependencies. I've pushed the change upstream and am patching this downstream just for today. Tomorrow going forward it should just be sourced from covid-data-model.